Bugfix and added quick argv:s for ease of use

+++ api/interfaces.go
@@ -63,7 +63,7 @@ type Inflection struct {
 
 type Suggestion struct {
 	Word      Word
-	Relevance int
+	Relevance float32
 }
 
 type SearchResponse struct {
